+++ packages/core/styles/src/packages/atomic/link/link.style.css.ts
@@ -17,6 +17,10 @@ const base = vanillaStyle({
 
   transition: `all ${varsThemeBase.motion.speed.fast} ease`,
 
+  backgroundColor: "transparent",
+  border: 0,
+  padding: 0,
+
   ":focus": {
     boxShadow: varsThemeBase.utils.focus,
   },