+++ packages/react/src/atomic/Button/src/button.stories.tsx
@@ -11,7 +11,10 @@ import { Button } from "./Button";
 export default {
   title: "Atomic/Button",
   component: Button,
-  subcomponents: { "Button.Skeleton": Button.Skeleton },
+  subcomponents: {
+    "Button.Skeleton": Button.Skeleton,
+    "Button.Link": Button.Link,
+  },
   argTypes: {
     children: { control: { disable: true } },
   },
@@ -23,6 +26,9 @@ export default {
 const Template: ComponentStory<typeof Button> = (args) => <Button {...args} />;
 const SkeletonTemplate: ComponentStory<typeof Button.Skeleton> = (args) => (
   <Button.Skeleton {...args} />
+);
+const LinkTemplate: ComponentStory<typeof Button.Link> = (args) => (
+  <Button.Link {...args} />
 );
 
 export const base = Template.bind({});
@@ -95,3 +101,10 @@ transparent.args = {
 
 export const skeleton = SkeletonTemplate.bind({});
 skeleton.args = {};
+
+export const asLink = LinkTemplate.bind({});
+asLink.args = {
+  children: "Button as anchor",
+  href: "https://nimbus.tiendanube.com/",
+  target: "_blank",
+};