+++ packages/react/src/atomic/Button/src/components/ButtonLink/index.ts
@@ -1,0 +1,4 @@
+import { ButtonLink } from "./ButtonLink";
+
+export { ButtonLink } from "./ButtonLink";
+export default ButtonLink;