+++ packages/react/src/atomic/Link/src/components/LinkButton/index.ts
@@ -1,0 +1,4 @@
+import { LinkButton } from "./LinkButton";
+
+export { LinkButton } from "./LinkButton";
+export default LinkButton;