+++ packages/react/src/atomic/Link/src/components/LinkButton/linkButton.types.ts
@@ -1,0 +1,5 @@
+import { ButtonHTMLAttributes } from "react";
+import { LinkProps } from "../../link.types";
+
+export type LinkButtonProps = Omit<LinkProps, "LinkHTMLAttributes"> &
+  ButtonHTMLAttributes<HTMLButtonElement>;