+++ packages/react/src/atomic/Link/src/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./LinkSkeleton";
+export * from "./LinkButton";